% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllMethods.R
\docType{class}
\name{FASTA-class}
\alias{FASTA-class}
\alias{FASTA}
\alias{show,FASTA-method}
\title{Class FASTA}
\usage{
FASTA(path)

\S4method{show}{FASTA}(object)
}
\arguments{
\item{path}{a path to a FASTA file}

\item{object}{an S4 class obect}
}
\description{
Class \code{FASTA} defines a FASTA file

This function provides a wrapper to create \code{FASTA} class
   objects based on a path to a FASTA file.
}
